+++ src/containers/Form/Users/Create/index.js
@@ -6,10 +6,14 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { reduxForm } from 'redux-form/immutable';
 import UsersForm from './layout';
-import { formValidation } from './validation';
-import { getSelectedEntityId, isCreating, userHasUpdatePermission } from '../../../../redux/modules/entities/selectors';
+import {
+  getSelectedEntityId,
+  isCreating,
+  userHasUpdatePermission
+} from '../../../../redux/modules/entities/selectors';
+import { formValidation } from './validation'
 import { onFormSubmit } from '../../../../redux/modules/entities';
-import { selectFormInitialValues, createFormName } from '../../../../redux/modules/form/selectors';
+import { selectCreateUserFormInitialValues, createFormName } from '../../../../redux/modules/form/selectors';
 import { selectTenantRoles, selectPlatformRoles } from '../../../../redux/modules/entities/roles/selectors';
 import { selectTenantIdentityProviders } from '../../../../redux/modules/entities/identityProviders/selectors';
 
@@ -19,8 +23,10 @@ const CreateUsersForm = compose(
     onSubmit: (values, dispatch, props) => {
       return dispatch(onFormSubmit(values, props));
     },
-    validate: formValidation,
-    destroyOnUnmount: true
+    enableReinitialize: true,
+    keepDirtyOnReinitialize: true,
+    destroyOnUnmount: true,
+    validate: formValidation
   })
 )(UsersForm);
 
@@ -29,7 +35,7 @@ export function mapStateToProps(state) {
     platformRoles: selectPlatformRoles(state),
     tenantRoles: selectTenantRoles(state),
     tenantIdentityProviders: selectTenantIdentityProviders(state),
-    initialValues: selectFormInitialValues(state),
+    initialValues: selectCreateUserFormInitialValues(state),
     isSaving: isCreating(state),
     userHasUpdatePermission: userHasUpdatePermission(state),
     key: getSelectedEntityId(state)