+++ src/containers/Form/Users/__tests__/index.test.js
@@ -18,13 +18,16 @@ import { getUserTenantStatus, getInvitationScenario } from '../../../../redux/mo
 import { selectTenantRoles, selectPlatformRoles } from '../../../../redux/modules/entities/roles/selectors';
 import { selectTenantIdentityProviders } from '../../../../redux/modules/entities/identityProviders/selectors';
 import { getCapacityRules } from '../../../../redux/modules/entities/capacityRules/selectors';
+import { getCurrentAgentId } from '../../../../redux/modules/userData/selectors';
 
 jest.mock('../../../../redux/modules/entities/identityProviders/selectors');
+jest.mock('../../../../redux/modules/userData/selectors');
 jest.mock('../../../../redux/modules/entities/capacityRules/selectors');
 selectTenantIdentityProviders.mockImplementation(() => []);
 
 jest.mock('../../../../redux/modules/entities/selectors');
 jest.mock('../../../../redux/modules/form/selectors');
+getCurrentAgentId.mockImplementation(() => '0000-0000-0000-0000');
 getCurrentForm.mockImplementation(() => 'gets form from state');
 getSelectedEntityId.mockImplementation(() => 'mockId');
 getCapacityRules.mockImplementation(() => [{ label: 'mockLabel', value: 'mockValue' }]);