+++ src/redux/modules/form/selectors.js
@@ -11,6 +11,7 @@ import {
   getSelectedEntityFormId
 } from '../entities/selectors';
 import { onFormSubmit, onSubEntityFormSubmit } from '../entities';
+import { selectFirstTenantRoleValue, selectFirstPlatformRoleValue } from '../entities/roles/selectors';
 
 import { entitiesMetaData } from '../entities/metaData';
 
@@ -51,6 +52,14 @@ export const selectFormInitialValues = state => {
       .delete(entitiesMetaData[getCurrentEntity(state)].dependentEntity);
   }
 };
+
+export const selectCreateUserFormInitialValues = state => {
+  return new Map({ 
+    inviteNow: true,
+    roleId: selectFirstTenantRoleValue(state),
+    platformRoleId: selectFirstPlatformRoleValue(state),
+  });
+}
 
 export const formSubmission = (values, dispatch, props) => dispatch(onFormSubmit(values, props));